+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## 0.8.0 (2023-09-26)
 
 - added popup response that updates the form when modifying related objects via the add and edit buttons
 - add init event to enable overwriting TomSelect parameters